# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: ZuzuFlow API
  version: "1.0.0"
  description: |
    REST API for ZuzuFlow.
    All endpoints are under `/api`. Executions broadcast real-time events
    over WebSocket at `ws://<host>/ws/executions?id=<executionId>`.
servers:
  - url: http://localhost:4000/api
    description: Local development
...
components:
  schemas:
    WorkflowStatus:
      type: string
      enum: [draft, active, inactive, archived]
...
    WorkflowTemplate:
      type: object
      required: [version, nodes, edges]
      properties:
        version:
          type: string
          example: "1.0"
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowEdge"
    WorkflowNode:
      type: object
      required: [id, kind, label, config]
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum:
            - manual
            - webhook
            - cron
            - mqtt_trigger
            - if_else
            - switch
            - delay
            - merge
            - http_request
            - send_email
            - postgres_query
            - custom_code
            - debug
          example: http_request
        label:
          type: string
          example: "Fetch User Data"
        config:
          type: object
          description: Node-specific configuration. Shape depends on `kind`.
          example: { url: "https://api.example.com/users", method: "GET" }
        position:
          type: object
          properties:
            x: { type: number }
            y: { type: number }
    WorkflowEdge:
      type: object
      required: [id, source, target]
      properties:
        id:
          type: string
          format: uuid
        source:
          type: string
          description: Source node ID
        target:
          type: string
          description: Target node ID
        sourceHandle:
          type: string
          description: Handle name (e.g. "true", "false" for If/Else; case values for Switch)
        label:
          type: string
    Workflow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Daily Report Generator"
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/WorkflowStatus"
        template:
          $ref: "#/components/schemas/WorkflowTemplate"
        version:
          type: integer
          description: Incremented on each update
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkflowListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/WorkflowStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkflowList:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowListItem"
...
    Error:
      type: object
      properties:
        error:
          type: string

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /workflows:
    get:
      tags: [Workflows]
      summary: List workflows
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/WorkflowStatus"
          description: Filter by status
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated list of workflows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowList"

    post:
      tags: [Workflows]
      summary: Create a workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, template]
              properties:
                name:
                  type: string
                  example: "My Workflow"
                description:
                  type: string
                template:
                  $ref: "#/components/schemas/WorkflowTemplate"
      responses:
        "201":
          description: Workflow created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "422":
          description: Invalid template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /workflows/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Workflows]
      summary: Get a workflow (includes template)
      responses:
        "200":
          description: Workflow detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "404":
          description: Not found

    put:
      tags: [Workflows]
      summary: Update a workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                template:
                  $ref: "#/components/schemas/WorkflowTemplate"
      responses:
        "200":
          description: Updated workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"

    delete:
      tags: [Workflows]
      summary: Delete a workflow
      responses:
        "204":
          description: Deleted
  /workflows/{id}/activate:
    post:
      tags: [Workflows]
      summary: Activate a workflow
      description: |
        Sets status to `active`. If the workflow contains a Cron trigger node,
        a Temporal Schedule is created so it fires automatically.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Activated workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{id}/deactivate:
    post:
      tags: [Workflows]
      summary: Deactivate a workflow
      description: |
        Sets status to `inactive`. If a Temporal Schedule exists (from a Cron trigger),
        it is deleted so no further automatic runs are scheduled.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deactivated workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"

  # ── Executions ───────────────────────────────────────────────────────────────
...